Changed all the tabs to each be their own form, so we can encapsulate the ui elements better.
Got main tab page functioning the way I want it to.  Figured out how to use drag and drop events to control the behavior.

Added ui elements for controlling groups and tied it to the same model as the infinite stash view since they should both always be showing the same thing.